[Test] Test reorganisation (#898)
Start reorganising the tests into more directories. The major divisions introduced are:

- `unit` for tests which at most require a `Mock` model
- `model_integration` for tests which require a 'real' model, but should work with any model
- `model_specific` for isolating issues we've found with particular LLMs
- `need_credentials` .... exactly what it says. Mainly remote-endpoint tests which don't have full Guidance support
- `notebooks` again, exactly what it says
- `server` contains tests of the local server. Technically `unit` tests, but run slowly

The workflows have been updated to match.
