+++ .github/workflows/action_server_tests.yml
@@ -25,12 +25,11 @@ jobs:
         run: |
           python -m pip install --upgrade pip
           pip install pytest
-          pip install -e .[all,test,bench]
+          pip install -e .[all,test]
           if [ -f requirements.txt ]; then pip install -r requirements.txt; fi
       - name: Run server tests
         shell: bash
         run: |
           pytest --cov=guidance --cov-report=xml --cov-report=term-missing \
-          -m server \
-          ./tests/
+          ./tests/server
