+++ tests/model_integration/library/test_gen.py
@@ -5,25 +5,7 @@ import pytest
 from guidance import gen, models, select
 
 
-def test_basic():
-    lm = models.Mock()
-    lm += "Write a number: " + gen("text", max_tokens=3)
-    assert len(lm["text"]) > 0
-
-
-def test_stop_string():
-    lm = models.Mock(b"<s>Count to 10: 1, 2, 3, 4, 5, 6, 7, 8, 9, 10")
-    lm += "Count to 10: 1, 2, 3, 4, 5, 6, 7, " + gen("text", stop=", 9")
-    assert lm["text"] == "8"
-
-
-def test_stop_char():
-    lm = models.Mock(b"<s>Count to 10: 1, 2, 3, 4, 5, 6, 7, 8, 9, 10")
-    lm += "Count to 10: 1, 2, 3, 4, 5, 6, 7, " + gen("text", stop=",")
-    assert lm["text"] == "8"
-
-
-def test_stop_list_side_effect(selected_model):
+def test_stop_list_side_effect(selected_model: models.Model):
     """Tests a bug where a stop list has an item appended to it in place instead of being updated non-destructively. The bug only occurs whe regex is also None"""
     stop_list = ["\nStep", "\n\n", "\nAnswer"]
     stop_list_length = len(stop_list)
@@ -57,14 +39,6 @@ Step 1"""
         + "\n"
     )
     assert stop_list_length == len(stop_list)
-
-
-def test_save_stop():
-    lm = models.Mock(b"<s>Count to 10: 1, 2, 3, 4, 5, 6, 7, 8, 9, 10")
-    lm += "Count to 10: 1, 2, 3, 4, 5, 6, 7, " + gen(
-        "text", stop=",", save_stop_text="stop_text"
-    )
-    assert lm["stop_text"] == ","
 
 
 def test_stop_quote(selected_model):
@@ -123,7 +97,7 @@ def test_metrics_select(selected_model: models.Model):
     )
 
 
-def test_unicode(selected_model):
+def test_unicode(selected_model: models.Model):
     # black makes this test ugly -- easier to read with fmt: off
     # fmt: off
     lm = selected_model
@@ -149,18 +123,7 @@ def test_unicode2(selected_model: models.Model):
     )
 
 
-def test_gsm8k():
-    lm = models.Mock()
-    (
-        lm
-        + """Question: Janet's ducks lay 16 eggs per day. She eats three for breakfast every morning and bakes muffins for her friends every day with four. She sells the remainder at the farmers' market daily for $2 per fresh duck egg. How much in dollars does she make every day at the farmers' market?
-Answer: """
-        + gen(max_tokens=30)
-    )
-    assert True
-
-
-def test_pattern_kleene(selected_model):
+def test_pattern_kleene(selected_model: models.Model):
     lm = selected_model
     lm += "The Lord is my"
     x = lm + gen(name="tmp", max_tokens=10)
@@ -171,75 +134,11 @@ def test_pattern_kleene(selected_model):
     )  # TODO: we just check startswith because exact token limits are not perfect yet...
 
 
-def test_non_token_force(selected_model):
+def test_non_token_force(selected_model: models.Model):
     """This forces some bytes that don't match a token (only longer tokens)"""
     lm = selected_model
     lm += "ae ae" + gen(regex=r"\d")
     assert len(str(lm)) == 6
-
-
-def test_pattern_optional():
-    lm = models.Mock(b"<s>12342333")
-    pattern = ".?233"
-    lm2 = lm + "123" + gen(name="numbers", regex=pattern, max_tokens=10)
-    assert lm2["numbers"] == "4233"
-    lm = models.Mock(b"<s>1232333")
-    pattern = ".?233"
-    lm2 = lm + "123" + gen(name="numbers", regex=pattern, max_tokens=10)
-    assert lm2["numbers"] == "233"
-    pattern = r"(Scott is bad)?(\d+)?o"
-    lm = models.Mock(b"<s>John was a little man full of things")
-    lm2 = lm + "J" + gen(name="test", regex=pattern, max_tokens=30)
-    assert lm2["test"] == "o"
-
-
-def test_pattern_stops_when_fulfilled():
-    lm = models.Mock(b"<s>123abc")
-    lm += gen(regex=r"\d+", max_tokens=10, name="test")
-    assert lm["test"] == "123"
-
-
-def test_pattern_star():
-    # lm = models.Mock(b"<s>1234233234<s>") # commented out because it is not a valid test
-    # patterns = ['\d+233', '\d*233', '.+233', '.*233']
-    # for pattern in patterns:
-    #     lm2 = lm + '123' + gen(name='numbers', regex=pattern, max_tokens=10)
-    #     assert lm2['numbers'] == '4233'
-    lm = models.Mock(b"<s>123233")
-    patterns = [r"\d*233", ".*233"]
-    for pattern in patterns:
-        lm2 = lm + "123" + gen(name="numbers", regex=pattern, max_tokens=10)
-        assert lm2["numbers"].startswith("233")
-    pattern = ".*(\n|little)"
-    lm = models.Mock(b"<s>John was a little")
-    lm2 = lm + "J" + gen(name="test", regex=pattern, max_tokens=30)
-    assert lm2["test"].startswith("ohn was a little")
-    lm = models.Mock(b"<s>John was a litt\n")
-    lm2 = lm + "J" + gen(name="test", regex=pattern, max_tokens=30)
-    assert lm2["test"].startswith("ohn was a litt\n")
-
-
-def test_stop_regex():
-    lm = models.Mock(b"<s>123a3233")
-    lm2 = lm + "123" + gen(name="test", stop_regex=r"\d233", max_tokens=10)
-    assert lm2["test"] == "a"
-    lm = models.Mock(b"<s>123aegalera3233")
-    lm2 = lm + "123" + gen(name="test", stop_regex=r"\d", max_tokens=30)
-    assert lm2["test"] == "aegalera"
-
-
-def test_stop_regex_star():
-    lm = models.Mock(b"<s>123a3233")
-    pattern = r"\d+233"
-    lm2 = lm + "123" + gen(name="test", stop_regex=pattern, max_tokens=10)
-    assert lm2["test"] == "a"
-
-
-def test_empty_pattern():
-    pattern = r"(Scott is bad)?(\d+)?"
-    lm = models.Mock(b"<s>J<s>")
-    lm2 = lm + "J" + gen(name="test", regex=pattern, max_tokens=30)
-    assert lm2["test"] == ""
 
 
 @pytest.mark.parametrize(
@@ -268,7 +167,7 @@ def test_various_regexes(selected_model: models.Model, prompt: str, pattern: str
     assert re.match(pattern, lm2["test"], re.DOTALL) is not None
 
 
-def test_long_prompt(selected_model, selected_model_name):
+def test_long_prompt(selected_model: models.Model, selected_model_name: str):
     if selected_model_name in [
         "hfllama7b",
         "hfllama_7b_gpu",
@@ -325,44 +224,7 @@ Step 1"""
     assert True
 
 
-def test_list_append():
-    """This tests is list append works across grammar appends."""
-    lm = models.Mock(b"<s>bababababa")
-    lm += "<s>"
-    for _ in range(3):
-        lm += gen("my_list", list_append=True, stop="a") + "a"
-    assert isinstance(lm["my_list"], list)
-    assert len(lm["my_list"]) == 3
-
-
-def test_list_append_in_grammar():
-    """This tests is list append works within the same grammar."""
-    lm = models.Mock(b"<s>bababababa")
-    lm += "<s>"
-    lm += (
-        gen("my_list", list_append=True, stop="a")
-        + "a"
-        + gen("my_list", list_append=True, stop="a")
-        + "a"
-        + gen("my_list", list_append=True, stop="a")
-    )
-    assert isinstance(lm["my_list"], list)
-    assert len(lm["my_list"]) == 3
-
-
-def test_one_char_suffix_and_regex():
-    model = models.Mock(b"<s>this is\na test")
-    model += gen(regex=".*", suffix="\n", max_tokens=20)
-    assert str(model) == "this is\n"
-
-
-def test_one_char_stop_and_regex():
-    model = models.Mock(b"<s>this is\na test")
-    model += gen(regex=".*", stop="\n", max_tokens=20)
-    assert str(model) == "this is"
-
-
-def test_tool_call(selected_model):
+def test_tool_call(selected_model: models.Model):
     import guidance
     from guidance import Tool, capture, one_or_more, select, zero_or_more
