+++ tests/model_specific/common_chat_testing.py
