+++ tests/need_credentials/test_azure_guidance.py
