+++ tests/need_credentials/test_lite_llm.py
