+++ tests/unit/library/test_pydantic.py
@@ -9,7 +9,7 @@ from pydantic.json_schema import to_jsonable_python as pydantic_to_jsonable_pyth
 from guidance import json as gen_json
 from guidance import models
 from guidance._grammar import Byte, ByteRange
-from ..utils import check_match_failure as _check_match_failure
+from ...utils import check_match_failure as _check_match_failure
 
 
 def to_compact_json(target: Any) -> str:
